langgraph-debate/requirements.txt
langgraph
langchain-openai
python-dotenv
langchain-community
